# Conflicts:
#	Linux/CodeBlocks/QuakeSpasm.cbp
#	MacOSX/English.lproj/InfoPlist.strings
#	MacOSX/Info.plist
#	MacOSX/QuakeSpasm.xcodeproj/project.pbxproj
#	MacOSX/QuakeSpasmPPC.xcodeproj/project.pbxproj
#	MacOSX/SDL.framework/Versions/A/SDL
#	MacOSX/SDL2.framework/Versions/A/Headers/SDL_assert.h
#	MacOSX/SDL2.framework/Versions/A/Headers/SDL_config_macosx.h
#	MacOSX/SDL2.framework/Versions/A/Headers/SDL_endian.h
#	MacOSX/SDL2.framework/Versions/A/Headers/SDL_rect.h
#	MacOSX/SDL2.framework/Versions/A/Headers/SDL_revision.h
#	MacOSX/SDL2.framework/Versions/A/Headers/SDL_stdinc.h
#	MacOSX/SDL2.framework/Versions/A/Headers/SDL_system.h
#	MacOSX/SDL2.framework/Versions/A/Headers/SDL_thread.h
#	MacOSX/SDL2.framework/Versions/A/SDL2
#	MacOSX/codecs/lib/libxmp.dylib
#	Quake/Makefile.darwin
#	Quake/Makefile.w32
#	Quake/Makefile.w64
#	Quake/common.h
#	Quake/gl_draw.c
#	Quake/gl_fog.c
#	Quake/gl_rmain.c
#	Quake/gl_sky.c
#	Quake/gl_texmgr.c
#	Quake/host.c
#	Quake/host_cmd.c
#	Quake/main_sdl.c
#	Quake/menu.c
#	Quake/protocol.h
#	Quake/r_alias.c
#	Quake/r_brush.c
#	Windows/SDL2/lib/SDL2.dll
#	Windows/SDL2/lib64/SDL2.dll
#	Windows/VisualStudio/quakespasm-sdl2.vcproj
#	Windows/VisualStudio/quakespasm.vcproj

+++ Quake/gl_model.c
@@ -532,7 +532,7 @@ void Mod_LoadTextures (lump_t *l)
 		if (((byte*)(mt+1) + pixels) > (mod_base + l->fileofs + l->filelen))
 		{
 			Con_DPrintf("Texture %s extends past end of lump\n", mt->name);
-			pixels = q_max(0, (mod_base + l->fileofs + l->filelen) - (byte*)(mt+1));
+			pixels = q_max(0L, (long)((mod_base + l->fileofs + l->filelen) - (byte*)(mt+1)));
 		}
 
 		tx->fullbright = NULL; //johnfitz
